+++ src/Components/JoinKeyword.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace PhpMyAdmin\SqlParser\Components;
 
 use PhpMyAdmin\SqlParser\Component;
-use PhpMyAdmin\SqlParser\Components\Parsers\Conditions;
+use PhpMyAdmin\SqlParser\Parsers\Conditions;
 
 use function array_search;
