+++ src/Parsers/ExpressionArray.php
@@ -2,54 +2,52 @@
 
 declare(strict_types=1);
 
-namespace PhpMyAdmin\SqlParser\Components\Parsers;
+namespace PhpMyAdmin\SqlParser\Parsers;
 
+use PhpMyAdmin\SqlParser\Components\CaseExpression;
 use PhpMyAdmin\SqlParser\Components\Expression;
-use PhpMyAdmin\SqlParser\Components\SetOperation;
+use PhpMyAdmin\SqlParser\Exceptions\ParserException;
 use PhpMyAdmin\SqlParser\Parseable;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Token;
 use PhpMyAdmin\SqlParser\TokensList;
 use PhpMyAdmin\SqlParser\TokenType;
 
+use function count;
 use function implode;
-use function in_array;
-use function trim;
+use function preg_match;
+use function strlen;
+use function substr;
 
 /**
- * `SET` keyword parser.
+ * Parses a list of expressions delimited by a comma.
  */
-final class SetOperations implements Parseable
+final class ExpressionArray implements Parseable
 {
     /**
      * @param Parser               $parser  the parser that serves as context
      * @param TokensList           $list    the list of tokens that are being parsed
      * @param array<string, mixed> $options parameters for parsing
      *
-     * @return SetOperation[]
+     * @return Expression[]
+     *
+     * @throws ParserException
      */
     public static function parse(Parser $parser, TokensList $list, array $options = []): array
     {
         $ret = [];
-
-        $expr = new SetOperation();
 
         /**
          * The state of the parser.
          *
          * Below are the states of the parser.
          *
-         *      0 ---------------------[ col_name ]--------------------> 0
-         *      0 ---------------------[ = or := ]---------------------> 1
-         *      1 -----------------------[ value ]---------------------> 1
+         *      0 ----------------------[ array ]---------------------> 1
+         *
          *      1 ------------------------[ , ]------------------------> 0
+         *      1 -----------------------[ else ]----------------------> (END)
          */
         $state = 0;
-
-        /**
-         * Token when the parser has seen the latest comma
-         */
-        $commaLastSeenAt = null;
 
         for (; $list->idx < $list->count; ++$list->idx) {
             /**
@@ -67,54 +65,59 @@ final class SetOperations implements Parseable
                 continue;
             }
 
-            // No keyword is expected.
             if (
                 ($token->type === TokenType::Keyword)
                 && ($token->flags & Token::FLAG_KEYWORD_RESERVED)
-                && ($state === 0)
+                && ((~$token->flags & Token::FLAG_KEYWORD_FUNCTION))
+                && ($token->value !== 'DUAL')
+                && ($token->value !== 'NULL')
+                && ($token->value !== 'CASE')
+                && ($token->value !== 'NOT')
             ) {
+                // No keyword is expected.
                 break;
             }
 
             if ($state === 0) {
-                if (in_array($token->token, ['=', ':='], true)) {
-                    $state = 1;
-                } elseif ($token->value !== ',') {
-                    $expr->column .= $token->token;
-                } elseif ($token->value === ',') {
-                    $commaLastSeenAt = $token;
+                if ($token->type === TokenType::Keyword && $token->value === 'CASE') {
+                    $expr = CaseExpression::parse($parser, $list, $options);
+                } else {
+                    $expr = Expression::parse($parser, $list, $options);
                 }
-            } else {
-                $tmp = Expression::parse(
-                    $parser,
-                    $list,
-                    ['breakOnAlias' => true],
-                );
-                if ($tmp === null) {
-                    $parser->error('Missing expression.', $token);
+
+                if ($expr === null) {
                     break;
                 }
 
-                $expr->column = trim($expr->column);
-                $expr->value = $tmp->expr;
                 $ret[] = $expr;
-                $expr = new SetOperation();
+                $state = 1;
+            } elseif ($state === 1) {
+                if ($token->value !== ',') {
+                    break;
+                }
+
                 $state = 0;
-                $commaLastSeenAt = null;
             }
         }
 
-        --$list->idx;
+        if ($state === 0) {
+            $parser->error('An expression was expected.', $list->tokens[$list->idx]);
+        }
 
-        // We saw a comma, but didn't see a column-value pair after it
-        if ($commaLastSeenAt !== null) {
-            $parser->error('Unexpected token.', $commaLastSeenAt);
+        --$list->idx;
+        $retIndex = count($ret) - 1;
+        if (isset($ret[$retIndex])) {
+            $expr = $ret[$retIndex]->expr;
+            if (preg_match('/\s*--\s.*$/', $expr, $matches)) {
+                $found = $matches[0];
+                $ret[$retIndex]->expr = substr($expr, 0, strlen($expr) - strlen($found));
+            }
         }
 
         return $ret;
     }
 
-    /** @param SetOperation[] $component the component to be built */
+    /** @param Expression[] $component the component to be built */
     public static function buildAll(array $component): string
     {
         return implode(', ', $component);
