+++ src/Statements/DeleteStatement.php
@@ -11,11 +11,11 @@ use PhpMyAdmin\SqlParser\Components\JoinKeyword;
 use PhpMyAdmin\SqlParser\Components\Limit;
 use PhpMyAdmin\SqlParser\Components\OptionsArray;
 use PhpMyAdmin\SqlParser\Components\OrderKeyword;
-use PhpMyAdmin\SqlParser\Components\Parsers\Conditions;
-use PhpMyAdmin\SqlParser\Components\Parsers\ExpressionArray;
-use PhpMyAdmin\SqlParser\Components\Parsers\JoinKeywords;
-use PhpMyAdmin\SqlParser\Components\Parsers\OrderKeywords;
 use PhpMyAdmin\SqlParser\Parser;
+use PhpMyAdmin\SqlParser\Parsers\Conditions;
+use PhpMyAdmin\SqlParser\Parsers\ExpressionArray;
+use PhpMyAdmin\SqlParser\Parsers\JoinKeywords;
+use PhpMyAdmin\SqlParser\Parsers\OrderKeywords;
 use PhpMyAdmin\SqlParser\Statement;
 use PhpMyAdmin\SqlParser\TokensList;
 use PhpMyAdmin\SqlParser\TokenType;
