+++ tests/Components/GroupKeywordTest.php
@@ -7,10 +7,9 @@ namespace PhpMyAdmin\SqlParser\Tests\Components;
 use Generator;
 use PhpMyAdmin\SqlParser\Components\Expression;
 use PhpMyAdmin\SqlParser\Components\GroupKeyword;
-use PhpMyAdmin\SqlParser\Components\Parsers\GroupKeywords;
+use PhpMyAdmin\SqlParser\Parsers\GroupKeywords;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 use PHPUnit\Framework\Attributes\DataProvider;
-
 use function array_map;
 use function is_array;
