+++ tests/Components/JoinKeywordTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\Parsers\JoinKeywords;
 use PhpMyAdmin\SqlParser\Parser;
+use PhpMyAdmin\SqlParser\Parsers\JoinKeywords;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class JoinKeywordTest extends TestCase
