+++ tests/Components/OrderKeywordTest.php
@@ -6,7 +6,7 @@ namespace PhpMyAdmin\SqlParser\Tests\Components;
 
 use PhpMyAdmin\SqlParser\Components\Expression;
 use PhpMyAdmin\SqlParser\Components\OrderKeyword;
-use PhpMyAdmin\SqlParser\Components\Parsers\OrderKeywords;
+use PhpMyAdmin\SqlParser\Parsers\OrderKeywords;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class OrderKeywordTest extends TestCase
