+++ tests/Components/RenameOperationTest.php
@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\Parsers\RenameOperations;
 use PhpMyAdmin\SqlParser\Parser;
+use PhpMyAdmin\SqlParser\Parsers\RenameOperations;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
 class RenameOperationTest extends TestCase
